Modify EnumRoles int for long. Modify method redirectIfNotAllowed in UserService, now he got role. Withe it, modify UserServlet. Modify CartServlet and LoginServlet withe config.properties. Added new value page.cart in config.properties

+++ src/main/java/ru/ncedu/ecomm/servlets/models/EnumRoles.java
@@ -7,13 +7,13 @@ public enum EnumRoles {
     MANAGER(2),
     USER(3);
 
-    private int role;
+    private long role;
 
-    EnumRoles(int role) {
+    EnumRoles(long role) {
         this.role = role;
     }
 
-    public int getRole() {
+    public long getRole() {
         return role;
     }
 }
